requirements.txt
timm
torch
torchvision
pandas
pillow
kaggle